requirements.txt
tensorflow==2.11.0
protobuf==3.19.6
streamlit==1.13.0
deepface==0.0.79
opencv-python==4.5.5.64
librosa==0.9.2
gTTS
playsound
SpeechRecognition
pydub
scikit-learn
altair==4.2.2
langchain==0.1.16
langchain-groq
python-dotenv
pyaudio

